Reorganised and refactored the code
