--- test/pageobjects/page.js
@@ -1,7 +1,0 @@
-const { browser } = require("@wdio/globals");
-
-module.exports = class Page {
-  open(path) {
-    return browser.url(`https://trello.com/${path}`);
-  }
-};
